[POO] aula08 ex1 refactoring
