--- 1ano/2semestre/poo/src/aula08/ex1/IEletricVehicle.java
@@ -1,6 +1,0 @@
-package aula08.ex1;
-
-public interface IEletricVehicle {
-    int currentBatteryLvl();
-    void charge(int percentage);
-}
