+++ 1ano/2semestre/poo/src/aula08/ex1/Interfaces/IElectricVehicle.java
@@ -1,0 +1,6 @@
+package aula08.ex1.Interfaces;
+
+public interface IElectricVehicle {
+    int currentBatteryLvl();
+    void charge(int percentage);
+}
