+++ 1ano/2semestre/poo/src/aula08/ex1/Interfaces/IFuelVehicle.java
@@ -1,4 +1,4 @@
-package aula08.ex1;
+package aula08.ex1.Interfaces;
 
 public interface IFuelVehicle {
     int fuelLevel();
