+++ 1ano/2semestre/poo/src/aula08/ex1/Main.java
@@ -1,5 +1,8 @@
 package aula08.ex1;
 
+import aula08.ex1.Enums.EngineType;
+import aula08.ex1.Vehicles.*;
+
 import java.util.Scanner;
 
 public class Main {
